Simplify and generalize navigation
