--- docs/_components/index.md
@@ -1,4 +1,0 @@
----
-permalink: /
-title: Components
----
