--- docs/_includes/nav/link-item.html
@@ -1,22 +1,0 @@
-{% assign is_current_page = include.link.href | eq: page.url -%}
-<a href="{{ include.link.href }}" class="{% if is_current_page %}usa-current{% endif %} usa-nav__link">
-  <span>
-    {{ include.link.text | smartify }}
-  </span>
-</a>
-{% unless include.flat -%}
-  {% assign is_anchor_link = include.link.href | starts_with: '#' -%}
-  {% if is_current_page or is_anchor_link -%}
-    {% if include.link.links -%}
-      <ul class="{{ include.ul_class }}">
-        {%
-          include nav/list.html
-          links=include.link.links
-          li_class=include.li_class
-          ul_class=include.ul_class
-          highlight_collection_when_is_current=include.highlight_collection_when_is_current
-        %}
-      </ul>
-    {% endif -%}
-  {% endif -%}
-{% endunless -%}
