--- docs/_includes/nav/list.html
@@ -1,26 +1,0 @@
-{% for link in include.links -%}
-  {% assign jekyll_page = link.href | resolve_permalink %}
-  <li class="{{ include.li_class }}">
-    {% if jekyll_page -%}
-      {%
-        include nav/page-item.html
-        jekyll_page=jekyll_page
-        flat=include.flat
-        links=link.links
-        li_class=include.li_class
-        ul_class=include.ul_class
-        highlight_collection_when_is_current=include.highlight_collection_when_is_current
-        text=link.text
-      %}
-    {% else -%}
-      {%
-        include nav/link-item.html
-        link=link
-        flat=include.flat
-        li_class=include.li_class
-        ul_class=include.ul_class
-        highlight_collection_when_is_current=include.highlight_collection_when_is_current
-      %}
-    {% endif %}
-  </li>
-{% endfor %}
