--- docs/_includes/nav/page-item.html
@@ -1,22 +1,0 @@
-{% assign is_current_page = true | if_is_current_page: include.jekyll_page, include.highlight_collection_when_is_current -%}
-<a href="{{ include.jekyll_page.url | relative_url }}" class="{% if is_current_page %}usa-current{% endif %} usa-nav__link">
-  <span>
-    {{ include.text | default: include.jekyll_page.title | smartify }}
-  </span>
-</a>
-{% unless include.flat -%}
-  {% if is_current_page -%}
-    {% assign links = include.links | default: include.jekyll_page.subnav -%}
-    {% if links -%}
-      <ul class="{{ include.ul_class }}">
-        {%
-          include nav/list.html
-          links=links
-          li_class=include.li_class
-          ul_class=include.ul_class
-          highlight_collection_when_is_current=include.highlight_collection_when_is_current
-        %}
-      </ul>
-    {% endif -%}
-  {% endif -%}
-{% endunless -%}
