--- docs/_plugins/generate_nav_for_components.rb
@@ -1,7 +1,0 @@
-require 'json'
-
-Jekyll::Hooks.register :site, :post_read do |site|
-  site.data['nav'] = site.collections['components'].docs.
-    filter { |doc| doc.basename != 'index.md'}.
-    map { |doc| { 'href' => doc.url } }
-end
