+++ docs/index.md
@@ -1,0 +1,7 @@
+---
+permalink: /
+title: Home
+layout: base
+---
+
+<meta http-equiv="refresh" content="0; url={{ site.components[0].url }}">
